# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EventManagement swagger application",
        "description": "Api d'une Application de  gestion des evenements ",
        "contact": {
            "email": "taneyemc@gmail.com"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/event/addEventType": {
            "post": {
                "tags": [
                    "event"
                ],
                "summary": "Add Event Type",
                "description": "Create a new event type",
                "operationId": "21dec565693de23f69731c738f29d649",
                "requestBody": {
                    "description": "Event type details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Event type name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Event type added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Event type added successfully"
                                        },
                                        "event_type": {
                                            "$ref": "#/components/schemas/EventType"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de validation"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "The name field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/event/updateEventStatus/{event}": {
            "put": {
                "tags": [
                    "event"
                ],
                "summary": "Update event status",
                "description": "Update an existing event by ID",
                "operationId": "3def4b1a2ee27e2d0e6d4711f614484d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the event",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Guest updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Event cancelled successfully"
                                        },
                                        "guest": {
                                            "$ref": "#/components/schemas/Guest"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Guest not found"
                    },
                    "422": {
                        "description": "The event is in progress cannot change the status."
                    },
                    "421": {
                        "description": "he event is already canceled."
                    }
                }
            }
        },
...
        "/api/event/showEvent/{event}": {
            "get": {
                "tags": [
                    "event"
                ],
                "summary": "Show event details",
                "description": "Retrieve details of a single event by its ID",
                "operationId": "5f399c45fa9869288584ba2f89c68c5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the event",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Event details retrieved successfully"
                                        },
                                        "event": {
                                            "$ref": "#/components/schemas/Event"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "EventType": {
                "required": [
                    "name",
                    "description"
                ],
                "properties": {
                    "name": {
                        "description": "Place's name",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Guest": {
                "required": [
                    "firstname",
                    "email",
                    "phone",
                    "place_id",
                    "table_id",
                    "event_id"
                ],
                "properties": {
                    "firstname": {
                        "description": "Guest's firstname",
                        "type": "string"
                    },
                    "lastname": {
                        "description": "Guest's lastname",
                        "type": "string"
                    },
                    "email": {
                        "description": "Guest's email",
                        "type": "string"
                    },
                    "gender": {
                        "description": "Guest's gender",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Guest's phone",
                        "type": "string"
                    },
                    "status": {
                        "description": "Guest's status",
                        "type": "enum"
                    },
                    "place_id": {
                        "description": "Guest's place_id",
                        "type": "string"
                    },
                    "table_id": {
                        "description": "Guest's table_id",
                        "type": "string"
                    },
                    "event_id": {
                        "description": "Guest's event_id",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}